requirements.txt
Flask
flask_cors
python-dotenv
flask_limiter
gunicorn
requests
